% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_election_results}
\alias{get_election_results}
\title{Get elections data for Kosovo}
\usage{
get_election_results(
  type = c("national", "local", "mayor", "mayor2"),
  year = NULL,
  by = c("party", "candidate"),
  level = c("national", "municipality", "pc", "ps")
)
}
\arguments{
\item{type}{c("national", "local", "mayor", "mayor2")}

\item{year}{From 2010 and above}

\item{by}{c("party", "candidate")}

\item{level}{c("national", "municipality", "pc", "ps")}
}
\value{

}
\description{
Get elections data for Kosovo
}
